// src/app/core/api/generated/model/catalog-field-binding-response.ts
/**
 * B4RRHH Personnel Administration API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CatalogFieldBindingResponse { 
    fieldCode: string;
    catalogKind: CatalogFieldBindingResponseCatalogKindEnum;
    ruleEntityTypeCode?: string | null;
    dependsOnFieldCode?: string | null;
    customResolverCode?: string | null;
    active: boolean;
}
export enum CatalogFieldBindingResponseCatalogKindEnum {
    Direct = 'DIRECT',
    Dependent = 'DEPENDENT',
    Custom = 'CUSTOM'
};